pyproject.toml
[project]
name = "qsit-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "python-docx>=1.2.0",
]
